Refactored Iterator implementation out of Observable into OperatorToIterator
- Trying to keep the massively large Observable.java class focused on the API and implementation details off in small classes.
