+++ rxjava-core/src/main/java/rx/Observable.java
@@ -1668,63 +1668,7 @@ public class Observable<T> {
      * @return the iterator that could be used to iterate over the elements of the observable.
      */
     public static <T> Iterator<T> getIterator(Observable<T> that) {
-        final BlockingQueue<Notification<T>> notifications = new LinkedBlockingQueue<Notification<T>>();
-
-        materialize(that).subscribe(new Observer<Notification<T>>() {
-            @Override
-            public void onCompleted() {
-                // ignore
-            }
-
-            @Override
-            public void onError(Exception e) {
-                // ignore
-            }
-
-            @Override
-            public void onNext(Notification<T> args) {
-                notifications.offer(args);
-            }
-        });
-
-        return new Iterator<T>() {
-            private Notification<T> buf;
-
-            @Override
-            public boolean hasNext() {
-                if (buf == null) {
-                    buf = take();
-                }
-                return !buf.isOnCompleted();
-            }
-
-            @Override
-            public T next() {
-                if (buf == null) {
-                    buf = take();
-                }
-                if (buf.isOnError()) {
-                    throw Exceptions.propagate(buf.getException());
-                }
-
-                T result = buf.getValue();
-                buf = null;
-                return result;
-            }
-
-            private Notification<T> take() {
-                try {
-                    return notifications.take();
-                } catch (InterruptedException e) {
-                    throw Exceptions.propagate(e);
-                }
-            }
-
-            @Override
-            public void remove() {
-                throw new UnsupportedOperationException("Read-only iterator");
-            }
-        };
+        return OperatorToIterator.toIterator(that);
     }
 
     /**
